Refactor: generalize metrics, integrating server and agents metrics API. Sender and qed server share the same metrics server.
