+++ gossip/processor.go
@@ -22,25 +22,25 @@ import (
 	"net/http"
 
 	"github.com/bbva/qed/log"
+	"github.com/bbva/qed/metrics"
 	"github.com/bbva/qed/protocol"
-	"github.com/prometheus/client_golang/prometheus"
 )
 
 type Processor interface {
 	Process(*protocol.BatchSnapshots)
-	RegisterMetrics(*prometheus.Registry)
+	RegisterMetrics(*metrics.Server)
 }
 
 type FakeProcessor struct {
 }
 
-func (d FakeProcessor) Process(b *protocol.BatchSnapshots)     {}
-func (d FakeProcessor) RegisterMetrics(r *prometheus.Registry) {}
+func (d FakeProcessor) Process(b *protocol.BatchSnapshots) {}
+func (d FakeProcessor) RegisterMetrics(m *metrics.Server)  {}
 
 type DummyProcessor struct {
 }
 
-func (d DummyProcessor) RegisterMetrics(r *prometheus.Registry) {}
+func (d DummyProcessor) RegisterMetrics(m *metrics.Server) {}
 
 func (d DummyProcessor) Process(b *protocol.BatchSnapshots) {
 	for i := 0; i < len(b.Snapshots); i++ {
